refactor: move token service implementation to infrastructure
