+++ src/infrastructure/adapters/services/TokenService.ts
@@ -1,8 +1,8 @@
 import jwt from 'jsonwebtoken'
-import { ITokenService } from './ITokenService'
-import { CodedError } from '../models/CodedError'
-import { TokenData } from '../models/TokenData'
-import { standardConfig } from '../config'
+import { ITokenService } from '../../../domain/services/ITokenService'
+import { standardConfig } from '../../../domain/config'
+import { CodedError } from '../../../domain/models/CodedError'
+import { TokenData } from '../../../domain/models/TokenData'
 
 /**
  * Token service
