+++ src/presentation/routes/AuthRoutes.ts
@@ -1,8 +1,8 @@
 import { Router } from 'express'
 import { AuthController } from '../controllers/AuthController'
-import { TokenService } from '../../domain/services/TokenService'
 import { TokenUseCase } from '../../application/usecases/TokenUseCase'
 import { TokenController } from '../controllers/TokenController'
+import { TokenService } from '../../infrastructure/adapters/services/TokenService'
 
 const router = Router()
